Updated package structure
