+++ src/test/java/com/dummy/api/tests/CreateUserTests.java
@@ -1,9 +1,10 @@
+package com.dummy.api.tests;
 
+import com.dummy.api.user.UserClient;
+import com.dummy.api.user.create.CreateUserRequestBody;
+import com.dummy.api.user.create.response.CreateUserResponse;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
-import users.UserClient;
-import users.create.CreateUserRequestBody;
-import users.create.response.CreateUserResponse;
 
 public class CreateUserTests {
