+++ src/test/java/com/dummy/api/user/UserClient.java
@@ -1,13 +1,11 @@
-package users;
+package com.dummy.api.user;
 
+import com.dummy.api.user.create.CreateUserRequestBody;
+import com.dummy.api.user.create.response.CreateUserErrorResponse;
+import com.dummy.api.user.create.response.CreateUserResponse;
+import com.dummy.api.user.get.GetUsersListResponse;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
-import users.create.CreateUserRequestBody;
-import users.create.response.CreateUserErrorResponse;
-import users.create.response.CreateUserResponse;
-import users.get.GetUsersListResponse;
-import users.post.PostCreateRequestBody;
-import users.post.response.PostCreateResponse;
 
 import static io.restassured.RestAssured.given;
 
@@ -45,20 +43,6 @@ public class UserClient {
         return response;
     }
 
-    public Response postCreate(PostCreateRequestBody body) {
-        Response response =
-                given()
-                        .contentType(ContentType.JSON)
-                        .header("app-id", "6305fad76cd510050784ce73")
-                        .body(body)
-                        .when()
-                        .post("https://dummyapi.io/data/v1/post/create");
-
-        response.then()
-                .log().body();
-        return response;
-    }
-
     public CreateUserResponse createUser(CreateUserRequestBody requestBody) {
         Response response = create(requestBody);
         CreateUserResponse createUserResponse = response.as(CreateUserResponse.class);
@@ -72,13 +56,4 @@ public class UserClient {
         errorResponse.setStatusCode(response.getStatusCode());
         return errorResponse;
     }
-
-    public PostCreateResponse createPost(PostCreateRequestBody postCreateRequestBody) {
-        Response response = postCreate(postCreateRequestBody);
-        PostCreateResponse postCreateResponse = response.as(PostCreateResponse.class);
-        postCreateResponse.setStatusCode(response.getStatusCode());
-        return postCreateResponse;
-    }
-
-
 }
